Rename wizard page component
