+++ apps/console/src/features/identity-providers/components/wizards/google/google-authentication-provider-create-wizard-content.tsx
